% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-GSmeta.R
\name{FEma}
\alias{FEma}
\title{A fixed effects (FE) meta-analysis (MA) of multiple GWAS with potentially dependent samples}
\usage{
FEma(betas, stders, sigma = NULL)
}
\arguments{
\item{betas}{effect size estimates from K studies}

\item{stders}{standard errors for the effct size estimates}

\item{sigma}{the correlation matrix of effect size estimates. Default to NULL for independent studies.}
}
\value{
\describe{
 \item{betam}{ estimated mean effect size }
 \item{Vm}{ variance of betam }
 \item{p.value}{ p-value for testing the mean effect size }
}
}
\description{
Under FE MA, we assume the same effect sizes across all studies. They can be efficiently estimated following the Lin-Sullivan approach.
}
\references{
Lin,D.Y. and Sullivan,P.F. (2009) Meta-Analysis of Genome-wide Association Studies with Overlapping Subjects. Am J Hum Genet 85, 862–872.

Han,B. and Eskin,E. (2011) Random-Effects Model Aimed at Discovering Associations in Meta-Analysis of Genome-wide Association Studies. The American Journal of Human Genetics 88, 586–598.

Wu,B. and Zhao,H. (2018) Powerful random effects modeling for meta-analysis of genome-wide association studies.
}
