% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE2C_HE.R
\name{RE2C}
\alias{RE2C}
\title{Improved random effects meta-analysis (RE2C) accounting for correlations and heterogeneous effects for GWAS}
\usage{
RE2C(beta, stders, sigma)
}
\arguments{
\item{beta}{effect size estimates from n studies}

\item{stders}{standard errors for the effct size estimates}

\item{sigma}{the correlation matrix of effect size estimates}
}
\value{
\describe{
 \item{RE2p}{ test p-value for overall effects (FE + RE) }
 \item{RE2Cp}{ test p-value for heterogenenous effect conditional on the FE }
 \item{LRT}{ LRT statistics for the FE and  RE in additiona to the FE  }
}
}
\description{
RE model accounts for fixed effects (FE) and heterogeneous effects in meta-analyzing potentially dependent GWAS. 
This implementation is based on the original software posted at http://software.buhmhan.com/RE2C with slight modification.
}
\references{
Lin,D.Y. and Sullivan,P.F. (2009) Meta-Analysis of Genome-wide Association Studies with Overlapping Subjects. Am J Hum Genet 85, 862–872.

Han,B. and Eskin,E. (2011) Random-Effects Model Aimed at Discovering Associations in Meta-Analysis of Genome-wide Association Studies. The American Journal of Human Genetics 88, 586–598.

Lee,C.H., Eskin,E., Han,B. (2017) Increasing the power of meta-analysis of genome-wide association studies to detect heterogeneous effects. Bioinformatics 33, i379–i388.
}
