% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-GSmeta.R
\name{REma}
\alias{REma}
\title{Random effects (RE) meta-analysis (MA) of multiple (potentially dependent) GWAS}
\usage{
REma(betas, Vb, B = 2500)
}
\arguments{
\item{betas}{effect size estimates from K studies}

\item{Vb}{covariance matrix of effect size estimates}

\item{B}{number of Monte Carlo samples to estimate the null distribution. Default to 2500. See Wu and Zhao (2018) reference.}
}
\value{
\describe{
 \item{p.value}{ p-values for FE (Pf), RE (Pr), RE conditional on FE (Pc) }
 \item{theta}{ proportion parameter in the chi-square mixture dist }
 \item{Q}{ LRT statistics for the FE (Qf) and RE (Qr) }
 \item{pars}{ estimated parameters (muf,mu,tau2) for FE mean, RE mean/variance parameters  }
}
}
\description{
Under RE MA, we assume heterogeneous effect sizes across all studies. This is potentially relevant
when meta-analyzing cross-disease GWAS with overlapped samples.
}
\examples{
K = 5; rho = 0.2
V = diag(K)*(1-rho) + rho
U0 = rnorm(K)*sqrt(1-rho) + rnorm(1)*sqrt(rho)
U1 = U0 + rnorm(K)*1.2
REma(U0,V)
REma(U1,V)
}
\references{
Lin,D.Y. and Sullivan,P.F. (2009) Meta-Analysis of Genome-wide Association Studies with Overlapping Subjects. Am J Hum Genet 85, 862–872.

Han,B. and Eskin,E. (2011) Random-Effects Model Aimed at Discovering Associations in Meta-Analysis of Genome-wide Association Studies. The American Journal of Human Genetics 88, 586–598.

Lee,C.H., Eskin,E., Han,B. (2017) Increasing the power of meta-analysis of genome-wide association studies to detect heterogeneous effects. Bioinformatics 33, i379–i388.

Wu,B. and Zhao,H. (2018) Powerful random effects modeling for meta-analysis of genome-wide association studies.
}
