% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-GSmeta.R
\name{REma}
\alias{REma}
\title{Random effects (RE) meta-analysis (MA) of multiple (potentially dependent) GWAS}
\usage{
REma(betas, Vb, B = 2500)
}
\arguments{
\item{betas}{effect size estimates from K studies}

\item{Vb}{covariance matrix of effect size estimates}

\item{B}{number of Monte Carlo samples to estimate the null distribution. Default to 2500. See Wu (2018) reference.}
}
\value{
\describe{
 \item{p.value}{ p-values for FE, RE, RE conditional on FE }
 \item{theta}{ proportion parameter in the chi-square mixture dist }
 \item{Q}{ LRT statistics for the FE and RE }
 \item{pars}{ estimated parameters (muf,mu,tau2) for FE mean, RE mean/variance parameters  }
}
}
\description{
Under RE MA, we assume heterogeneous effect sizes across all studies. This is potentially relevant
when meta-analyzing cross-disease GWAS with overlapped samples.
}
\references{
Lin,D.Y. and Sullivan,P.F. (2009) Meta-Analysis of Genome-wide Association Studies with Overlapping Subjects. Am J Hum Genet 85, 862–872.

Han,B. and Eskin,E. (2011) Random-Effects Model Aimed at Discovering Associations in Meta-Analysis of Genome-wide Association Studies. The American Journal of Human Genetics 88, 586–598.

Lee,C.H., Eskin,E., Han,B. (2017) Increasing the power of meta-analysis of genome-wide association studies to detect heterogeneous effects. Bioinformatics 33, i379–i388. 

Wu,B. (2018) On random effects modeling for meta-analysis of genome-wide associations studies.
}
